requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
requests==2.32.4
xmltodict==0.14.2
python-dotenv==1.1.1
reportlab==4.4.2
sqlalchemy==2.0.41
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.0
